+++ crates/gdcef/src/cef_texture/mod.rs
@@ -493,9 +493,9 @@ impl CefTexture {
             .audio_sample_rate
             .as_ref()
             .and_then(|sr| sr.lock().ok().map(|sr| *sr))
-            .unwrap_or(48000);
+            .unwrap_or(48000.0);
 
-        stream.set_mix_rate(sample_rate as f32);
+        stream.set_mix_rate(sample_rate);
         stream.set_buffer_length(0.1);
 
         stream
